Introduce new folder structure
